finished getAccounts with optional MaxResult QueryParameter

+++ models/database.go
@@ -2,7 +2,7 @@ package models
 
 type Database interface {
 	CreateAccount(req *AccountRequest) (*Account, error)
-	GetAllAccounts(maxResult int) ([]*Account, error)
+	GetAllAccounts(maxResult uint64) ([]*Account, error)
 	GetAccountByID(id string) (*Account, error)
 	DeleteAccount(id string) (*Account, error)
 	UpdateAccount(id string, req *AccountRequest) (*Account, error)
